+++ resources/js/ssr.tsx
@@ -1,6 +1,7 @@
 import { createInertiaApp } from '@inertiajs/react';
 import createServer from '@inertiajs/react/server';
 import { configureEcho } from '@laravel/echo-react';
+import type { ComponentType } from 'react';
 import ReactDOMServer from 'react-dom/server';
 
 const appName = import.meta.env.VITE_APP_NAME || 'Laravel';
@@ -14,8 +15,8 @@ createServer((page) =>
         render: ReactDOMServer.renderToString,
         title: (title) => (title ? `${title} - ${appName}` : appName),
         resolve: (name) => {
-            const pages = import.meta.glob('./pages/**/*.tsx');
-            return pages[`./pages/${name}.tsx`]();
+            const pages = import.meta.glob<{ default: ComponentType }>('./pages/**/*.tsx', { eager: true });
+            return pages[`./pages/${name}.tsx`];
         },
         setup: ({ App, props }) => {
             return <App {...props} />;